Update UpdateX3ConfigFilesForServerNameChange.ps1
Set-config was setting the DB_NAM incorrectly. Collapsing the white space around the equals sign allows the DB_NAM to be set correctly within a batch file. If it was set incorrectly you'd receive an error on a folder copy within the Admin console. It would tell you that the SQL Server could not be found when trying to run the E:\SAGE\SAGEX3V6\X3V6PRM\runtime\ebin\adcrapss7_std10.bat script.