Edit-XmlAttribute function added,
This checkin represents partial work. Edit-XmlAttribute was added. This
function is different than Edit-XmlNodes. The former alters an XML
attribute whereas the latter alters an XML node / element value.

Several places within the x3world.xml file and more need to be handled
in order for the web server to function properly after a rename. This
checkin is just starting that work.