Example Web Services Wrapper Debugging
This file is an example of how to build a wrapper debug for the SOH web
service object. If the ZZWSSOH.src file exists X3 will call this during
the execution of your web service. This file shows you how to set
variables for a line on the sales order grid.