Get Last Accessed Date From Files And Export to CSV
This procedure gets the last accessed date from a file on your OS. The
last accessed date in Windows is only accurate if the registry key
mentioned in the code is set to 0.

For Sage ERP X3 you can use this script to figure out who's running what
Sage ERP X3 reports and when.